requirements.txt
langchain-groq
Flask
requests
flask-cors
python-dotenv